Improvement: Move performance class to correct location
