+++ classes/local/performance/actions/performance_action_interface.php
@@ -22,7 +22,7 @@
  * @license http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace mod_booking\performance\actions;
+namespace mod_booking\local\performance\actions;
 
 defined('MOODLE_INTERNAL') || die();
